# openapi.yaml
openapi: 3.0.0
info:
  title: Baby Billing API
  version: 1.0.0
  description: API для использования триггеров по взаимодействию между микросервисами Cdr generator, Brt и Hrs, а также для управления клиентской и платежной информацией
servers:
  - url: http://localhost:8081/api
  - url: http://localhost:8082/api
paths:
...
  /subscriber/balance:
    get:
      summary: Получить информацию о балансе
      description: Этот метод позволяет получить информацию о балансе клиента "Ромашки". Использовать хост с портом "8082".
      parameters:
        - in: query
          name: subscriber
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
  /subscriber/deposit:
    post:
      summary: Депозит
      description: Этот метод позволяет сделать депозит на определенную сумму для существующего клиента "Ромашки". Использовать хост с портом "8082".
      parameters:
        - in: query
          name: subscriber
          required: true
          schema:
            type: string
        - in: query
          name: amount
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: string
                example: "Successfully made a deposit."
...
components:
  schemas:
...
    Balance:
      type: number
      format: double